refactor: Move RandomInstanceType to fixtures
